refactor(robot-server, system-server): commonize server code (#12113)
* Added shell of `server-utils` project
* Added server-utils to robot server's pipfile
* Added server-utils to system-server pipfile
* refactor(robot-server): moved `util.py` to server-utils
* refactor(robot-server): moved app_state.py to server-utils
* Added a server-utils github workflow

---------
