+++ .github/workflows/robot-server-lint-test.yaml
@@ -11,6 +11,7 @@ on:
       - 'hardware/**/*'
       - 'Makefile'
       - 'shared-data/**/*'
+      - 'server-utils/**/*'
       - '!shared-data/js/**/*'
       - 'robot-server/**/*'
       - 'notify-server/**/*'
@@ -30,6 +31,7 @@ on:
       - 'hardware/**/*'
       - 'Makefile'
       - 'shared-data/**/*'
+      - 'server-utils/**/*'
       - '!shared-data/js/**/*'
       - 'robot-server/**/*'
       - 'notify-server/**/*'
