+++ scripts/python_build_utils.py
@@ -32,6 +32,7 @@ package_entries = {
     'hardware': PackageEntry('opentrons_hardware'),
     'usb-bridge': PackageEntry('usb_bridge'),
     'system-server': PackageEntry('system_server'),
+    'server-utils': PackageEntry('server_utils'),
 }
 
 project_entries = {
