+++ server-utils/.flake8
@@ -1,0 +1,31 @@
+[flake8]
+
+# max cyclomatic complexity
+max-complexity = 9
+
+extend-ignore =
+    # defer formatting concerns to black
+    # E203: space around `:` operator
+    # E501: maximum line length
+    E203,
+    E501,
+    # do not require type annotations for self nor cls
+    ANN101,
+    ANN102
+    # do not require docstring for __init__, put them on the class
+    D107,
+
+# configure flake8-docstrings
+# https://pypi.org/project/flake8-docstrings/
+docstring-convention = google
+
+noqa-require-code = true
+
+# TODO(mc, 2021-04-12): this config ignores type annotation and doc
+# string lints in these modules; remove entries as they are fixed
+per-file-ignores =
+    setup.py:ANN,D
+    server_utils/util.py:ANN,D
+    server_utils/fastapi_utils/service/*:ANN,D
+    tests/fastapi_utils/service/*:ANN,D
+    tests/test_util.py:ANN,D
