+++ server-utils/.gitignore
@@ -1,0 +1,5 @@
+# This is generated by setuptools
+pyproject.toml
+
+# This is generated by sdist
+LICENSE
