+++ server-utils/Pipfile
@@ -7,7 +7,7 @@ name = "pypi"
 python_version = "3.7"
 
 [dev-packages]
-robot-server = { editable = true, path = "." }
+server-utils = { editable = true, path = "." }
 # pytest dependencies on windows, spec'd here to force lockfile inclusion
 # https://github.com/pypa/pipenv/issues/4408#issuecomment-668324177
 atomicwrites = { version = "==1.4.0", sys_platform = "== 'win32'" }
@@ -18,8 +18,6 @@ pytest-cov = "==2.10.1"
 pytest-lazy-fixture = "==0.6.3"
 pytest-xdist = "~=2.2.1"
 requests = "==2.26.0"
-tavern = "~=1.6"
-graphviz = "==0.17"
 mock = "~=4.0.2"
 mypy = "==0.910"
 flake8 = "~=3.9.0"
@@ -35,9 +33,6 @@ sqlalchemy2-stubs = "==0.0.2a21"
 python-box = "==5.4.1"
 
 [packages]
-opentrons = { editable = true, path = "../api" }
-opentrons-shared-data = { editable = true, path = "../shared-data/python" }
-notify-server = { editable = true, path = "../notify-server" }
 anyio = "==3.3.0"
 fastapi = "==0.68.1"
 python-dotenv = "==0.19.0"
@@ -48,13 +43,5 @@ uvicorn = "==0.14.0"
 wsproto = "==1.0.0"
 systemd-python = { version = "==234", sys_platform = "== 'linux'" }
 idna = "==3.2"
-click = "==8.0.1"
-# TODO(mc, 2022-04-04): this dep is newer than the OT-2 for dev environment reasons
-# remove this TODO when Opentrons/buildroot has updated to newer numpy
-# https://github.com/Opentrons/buildroot/pull/148
-numpy = "==1.21.2"
-zipp = "==3.5.0"
-opentrons-hardware = { editable = true, path = "./../hardware" }
 sqlalchemy = "==1.4.32"
 aiosqlite = "==0.17.0"
-server-utils = {editable = true, path = "./../server-utils"}
