+++ server-utils/mypy.ini
@@ -1,0 +1,19 @@
+[mypy]
+plugins = pydantic.mypy, decoy.mypy, sqlalchemy.ext.mypy.plugin
+show_error_codes = True
+strict = True
+
+[pydantic-mypy]
+init_forbid_extra = True
+init_typed = True
+warn_required_dynamic_aliases = True
+warn_untyped_fields = True
+
+
+# TODO(mc, 2021-09-08): fix and remove any / all of the
+# overrides below whenever able
+
+[mypy-tests.fastapi_utils.service.json_api.*]
+disallow_any_generics = False
+disallow_untyped_defs = False
+disallow_incomplete_defs = False
