+++ server-utils/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+addopts = --color=yes
+asyncio_mode = auto
