+++ server-utils/server_utils/__init__.py
@@ -1,0 +1,5 @@
+"""Opentrons HTTP server utils.
+
+This package provides common utilities to be shared across any Opentrons servers
+intended to run on Opentrons robots.
+"""
