+++ server-utils/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Opentrons server_utils package."""
