+++ system-server/Pipfile
@@ -15,6 +15,7 @@ pydantic = "==1.8.2"
 sqlalchemy = "==1.4.32"
 systemd-python = { version = "==234", sys_platform = "== 'linux'" }
 importlib-metadata = ">=4.13.0,<5"
+server-utils = {editable = true, path = "./../server-utils"}
 
 [dev-packages]
 system_server = {path = ".", editable = true}
