+++ system-server/system_server/persistence/__init__.py
@@ -12,7 +12,11 @@ from .tables import registration_table, migration_table
 from .persistent_directory import create_persistent_directory
 from .system_uuid import get_system_uuid
 
-from system_server.app_state import AppState, AppStateAccessor, get_app_state
+from server_utils.fastapi_utils.app_state import (
+    AppState,
+    AppStateAccessor,
+    get_app_state,
+)
 from system_server.settings import get_settings
 from system_server.connection import AuthorizationTracker
